+++ src/settings.rs
@@ -6,7 +6,6 @@ use crate::cmd::Command;
 pub struct Settings {
     pub venv_from_stdlib: bool,
     pub venv_outside_project: bool,
-    pub system_site_packages: bool,
     pub production: bool,
 }
 
@@ -15,7 +14,6 @@ impl Default for Settings {
         Settings {
             venv_from_stdlib: true,
             venv_outside_project: false,
-            system_site_packages: false,
             production: false,
         }
     }
@@ -23,14 +21,13 @@ impl Default for Settings {
 
 impl Settings {
     /// Construct a new Settings instance using
-    /// options fromm the command line (the `cmd` parameter)
-    /// and enviornment variables.
+    /// options from the command line (the `cmd` parameter)
+    /// and environment variables.
     //
     // Note:  Called in `main()` and in test heplers.
     pub fn from_shell(cmd: &Command) -> Settings {
         let mut res = Settings {
             production: cmd.production,
-            system_site_packages: cmd.system_site_packages,
             ..Default::default()
         };
         if std::env::var("DMENV_NO_VENV_STDLIB").is_ok() {